+++ lib/src/widgets/drag_targets/vertical_drag_target.dart
@@ -116,9 +116,10 @@ class _VerticalDragTargetState extends State<VerticalDragTarget> with SnapPoints
     super.initState();
 
     WidgetsBinding.instance.addPostFrameCallback((_) {
+      if (!mounted) return;
       _updateSnapPoints();
       widget.snapping.addListener(_updateSnapPoints);
-      controller.visibleEvents.addListener(_updateSnapPoints);
+      widget.viewController.visibleEvents.addListener(_updateSnapPoints);
     });
   }
 
@@ -129,16 +130,16 @@ class _VerticalDragTargetState extends State<VerticalDragTarget> with SnapPoints
       oldWidget.snapping.removeListener(_updateSnapPoints);
       widget.snapping.addListener(_updateSnapPoints);
     }
-    if (oldWidget.controller != widget.controller) {
-      oldWidget.controller.visibleEvents.removeListener(_updateSnapPoints);
-      widget.controller.visibleEvents.addListener(_updateSnapPoints);
+    if (oldWidget.viewController != widget.viewController) {
+      oldWidget.viewController.visibleEvents.removeListener(_updateSnapPoints);
+      widget.viewController.visibleEvents.addListener(_updateSnapPoints);
     }
   }
 
   @override
   void dispose() {
     widget.snapping.removeListener(_updateSnapPoints);
-    controller.visibleEvents.removeListener(_updateSnapPoints);
+    widget.viewController.visibleEvents.removeListener(_updateSnapPoints);
     super.dispose();
   }
 
@@ -146,7 +147,7 @@ class _VerticalDragTargetState extends State<VerticalDragTarget> with SnapPoints
   void _updateSnapPoints() {
     if (!snapping.snapToOtherEvents) return;
     clearSnapPoints();
-    addEventSnapPoints(controller.visibleEvents.value, context.location);
+    addEventSnapPoints(widget.viewController.visibleEvents.value, context.location);
   }
 
   @override