+++ lib/src/widgets/multi_day/multi_day_body.dart
@@ -246,12 +246,9 @@ class _MultiDayPageState extends State<MultiDayPage> {
                 end: visibleRange.start.add(Duration(days: widget.viewController.viewConfiguration.numberOfDays)),
               )
             : visibleRange;
-        final controller = context.kalenderController;
-        controller.floatingVisibleRange.value = range;
+        widget.viewController.floatingVisibleRange.value = range;
         _updateVisibleEvents(index, context.location);
-
-        final callbacks = context.callbacks;
-        callbacks?.onPageChanged?.call(controller.visibleDateTimeRange.value!);
+        context.callbacks?.onPageChanged?.call(range.forLocation(location: context.location));
       },
       itemBuilder: (context, index) {
         final visibleRange = _pageNavigation.rangeFromIndex(index, context.location);